# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
...
        "/films/images/poster/{film_id}": {
            "patch": {
                "tags": [
                    "Films",
                    "FilmImages",
                    "Unfinished"
                ],
                "summary": "Update Film Poster",
                "operationId": "update_film_poster_films_images_poster__film_id__patch",
                "security": [
                    {
                        "OAuth2PasswordBearer": [
                            "write:films"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "name": "film_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Film Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_update_film_poster_films_images_poster__film_id__patch"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/films/images/posters": {
            "patch": {
                "tags": [
                    "Films",
                    "FilmImages",
                    "Unfinished"
                ],
                "summary": "Add Film Posters",
                "operationId": "add_film_posters_films_images_posters_patch",
                "security": [
                    {
                        "OAuth2PasswordBearer": [
                            "write:films"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "name": "film_title",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Title of the film"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_add_film_posters_films_images_posters_patch"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Films",
                    "FilmImages",
                    "Unfinished"
                ],
                "summary": "Delete Film Posters",
                "operationId": "delete_film_posters_films_images_posters_delete",
                "security": [
                    {
                        "OAuth2PasswordBearer": [
                            "write:films"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "name": "file_name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "File Name"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/films/images/image/{file_name}": {
            "get": {
                "tags": [
                    "Films",
                    "FilmImages",
                    "Unfinished"
                ],
                "summary": "Get Image",
                "operationId": "get_image_films_images_image__file_name__get",
                "parameters": [
                    {
                        "name": "file_name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "File Name"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/films/film": {
            "post": {
                "tags": [
                    "Films",
                    "Unfinished"
                ],
                "summary": "Create Film",
                "operationId": "create_film_films_film_post",
                "security": [
                    {
                        "OAuth2PasswordBearer": [
                            "write:films"
                        ]
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Film"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Film"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Films",
                    "Unfinished"
                ],
                "summary": "Update Film",
                "operationId": "update_film_films_film_patch",
                "security": [
                    {
                        "OAuth2PasswordBearer": [
                            "write:films"
                        ]
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Film"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "Films",
                    "Unfinished"
                ],
                "summary": "Get Film",
                "operationId": "get_film_films_film_get",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Title"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Film"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/films/film/{film_id}": {
            "get": {
                "tags": [
                    "Films",
                    "Unfinished"
                ],
                "summary": "Get Film By Id",
                "operationId": "get_film_by_id_films_film__film_id__get",
                "parameters": [
                    {
                        "name": "film_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Film Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Film"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/films/films": {
            "get": {
                "tags": [
                    "Films",
                    "Unfinished"
                ],
                "summary": "Get Films",
                "operationId": "get_films_films_films_get",
                "parameters": [
                    {
                        "name": "start",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "title": "Range starting ID to get"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "title": "Amount of resources to fetch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Film"
                                    },
                                    "title": "Response Get Films Films Films Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Body_add_film_posters_films_images_posters_patch": {
                "properties": {
                    "files": {
                        "items": {
                            "type": "string",
                            "format": "binary"
                        },
                        "type": "array",
                        "title": "Files"
                    }
                },
                "type": "object",
                "required": [
                    "files"
                ],
                "title": "Body_add_film_posters_films_images_posters_patch"
            },
...
            "Body_update_film_poster_films_images_poster__film_id__patch": {
                "properties": {
                    "file": {
                        "type": "string",
                        "format": "binary",
                        "title": "File"
                    }
                },
                "type": "object",
                "required": [
                    "file"
                ],
                "title": "Body_update_film_poster_films_images_poster__film_id__patch"
            },
...
            "Film": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "age_rating": {
                        "type": "string",
                        "title": "Age Rating"
                    },
                    "duration_sec": {
                        "type": "integer",
                        "minimum": 1.0,
                        "title": "Duration Sec"
                    },
                    "trailer_desc": {
                        "type": "string",
                        "title": "Trailer Desc"
                    },
                    "on_air_from": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "string"
                            }
                        ],
                        "title": "On Air From"
                    },
                    "on_air_to": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "string"
                            }
                        ],
                        "title": "On Air To"
                    },
                    "is_active": {
                        "type": "boolean",
                        "title": "Is Active"
                    },
                    "images": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/FilmImage"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Images"
                    },
                    "schedules": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/Schedule"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Schedules"
                    },
                    "class_name": {
                        "type": "string",
                        "title": "Class Name",
                        "default": "FILM"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "title",
                    "age_rating",
                    "duration_sec",
                    "trailer_desc",
                    "on_air_from",
                    "on_air_to",
                    "is_active"
                ],
                "title": "Film"
            },
            "FilmImage": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "filename": {
                        "type": "string",
                        "title": "Filename"
                    },
                    "class_name": {
                        "type": "string",
                        "title": "Class Name",
                        "default": "FILM_IMAGE"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "filename"
                ],
                "title": "FilmImage"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "Schedule": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "hall_id": {
                        "type": "integer",
                        "title": "Hall Id"
                    },
                    "film_id": {
                        "type": "integer",
                        "title": "Film Id"
                    },
                    "show_time": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Show Time"
                    },
                    "on_schedule": {
                        "type": "boolean",
                        "title": "On Schedule"
                    },
                    "ticket_price": {
                        "type": "number",
                        "minimum": 1.0,
                        "title": "Ticket Price"
                    },
                    "class_name": {
                        "type": "string",
                        "title": "Class Name",
                        "default": "SCHEDULE"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "hall_id",
                    "film_id",
                    "show_time",
                    "on_schedule",
                    "ticket_price"
                ],
                "title": "Schedule"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "token"
                    }
                }
            }
        }
    }
}